+++ src/Candoumbe.Pipelines/Components/Formatting/DotNetFormatter.cs
@@ -1,5 +1,5 @@
-﻿using System.ComponentModel;
-using Nuke.Common.Tooling;
+using System.ComponentModel;
+using Fallout.Common.Tooling;
 
 namespace Candoumbe.Pipelines.Components.Formatting;