+++ src/Candoumbe.Pipelines/Components/GitHub/IPullRequest.cs
@@ -1,5 +1,5 @@
-﻿using System;
-using Nuke.Common;
+using System;
+using Fallout.Common;
 
 namespace Candoumbe.Pipelines.Components.GitHub;