+++ src/Candoumbe.Pipelines/Components/IHaveGitRepository.cs
@@ -1,12 +1,12 @@
-﻿using Nuke.Common;
-using Nuke.Common.Git;
+using Fallout.Common;
+using Fallout.Common.Git;
 
 namespace Candoumbe.Pipelines.Components;
 
 /// <summary>
 /// Marks a pipeline that has a git repository
 /// </summary>
-public interface IHaveGitRepository : INukeBuild
+public interface IHaveGitRepository : IFalloutBuild
 {
     /// <summary>
     /// The git repository of the project which a pipeline handles