+++ src/Candoumbe.Pipelines/Components/IHaveOutput.cs
@@ -1,12 +1,12 @@
-﻿using Nuke.Common;
-using Nuke.Common.IO;
+using Fallout.Common;
+using Fallout.Common.IO;
 
 namespace Candoumbe.Pipelines.Components;
 
 /// <summary>
 /// Defines an output directory
 /// </summary>
-public interface IHaveOutputDirectory : INukeBuild
+public interface IHaveOutputDirectory : IFalloutBuild
 {
     /// <summary>
     /// Name of the root directory that contain all output