+++ src/Candoumbe.Pipelines/Components/IHaveSolution.cs
@@ -1,7 +1,7 @@
-﻿using JetBrains.Annotations;
+using JetBrains.Annotations;
 
-using Nuke.Common;
-using Nuke.Common.ProjectModel;
+using Fallout.Common;
+using Fallout.Common.ProjectModel;
 
 namespace Candoumbe.Pipelines.Components;
 
@@ -9,7 +9,7 @@ namespace Candoumbe.Pipelines.Components;
 /// Marks a pipeline that handle a Solution
 /// </summary>
 [PublicAPI]
-public interface IHaveSolution : INukeBuild
+public interface IHaveSolution : IFalloutBuild
 {
     /// <summary>
     /// The solution to build